# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Public API documentation served through QRUPI API Gateway.",
        "title": "QRUPI V2 API",
        "contact": {},
        "version": "2.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/attendance-logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Attendance Log API"
                ],
                "summary": "Get all Attendance Logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sort by field",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "Sort order",
                        "name": "sortOrder",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of results per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Attendance Log API"
                ],
                "summary": "Create Attendance Log",
                "parameters": [
                    {
                        "description": "Attendance Log Data",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/attendance_logs.CreateDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/attendance-logs/check-in": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Attendance Log API"
                ],
                "summary": "Check in Attendance",
                "parameters": [
                    {
                        "description": "Attendance Check-in Data",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/attendance_logs.CheckInDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/attendance-logs/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Attendance Log API"
                ],
                "summary": "Get Attendance Log by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Attendance Log ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Attendance Log API"
                ],
                "summary": "Update Attendance Log",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Attendance Log ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Attendance Log Data",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/attendance_logs.UpdateDTO"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Attendance Log API"
                ],
                "summary": "Delete Attendance Log permanently",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Attendance Log ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/attendance-logs/{id}/archived": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Attendance Log API"
                ],
                "summary": "Archive Attendance Log",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Attendance Log ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/attendance-logs/{id}/check-out": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Attendance Log API"
                ],
                "summary": "Check out Attendance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Attendance Log ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Attendance Check-out Data",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/attendance_logs.CheckOutDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth API"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "Login Data",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth API"
                ],
                "summary": "Register user",
                "parameters": [
                    {
                        "description": "Register Data",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RegisterDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/auth/reset-password": {
            "post": {
                "tags": [
                    "Auth API"
                ],
                "summary": "Reset password user",
                "parameters": [
                    {
                        "description": "Reset Password Data",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.ResetPasswordDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/auth/student/scan": {
            "post": {
                "tags": [
                    "Auth API"
                ],
                "summary": "Login student using barcode and PIN",
                "parameters": [
                    {
                        "description": "Student Scan Login Data",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.StudentScanLoginDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/export/excel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Export API"
                ],
                "summary": "Export excel",
                "parameters": [
                    {
                        "description": "Export Data",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/export.ExportDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "attendance_logs.CheckInDTO": {
            "type": "object",
            "required": [
                "status",
                "type",
                "user_id"
            ],
            "properties": {
                "check_in_at": {
                    "type": "string"
                },
                "learning_group_id": {
                    "type": "string"
                },
                "location_lat": {
                    "type": "number"
                },
                "location_long": {
                    "type": "number"
                },
                "requires_check_out": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "on_time",
                        "late"
                    ]
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "student",
                        "teacher",
                        "learning_group"
                    ]
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "attendance_logs.CheckOutDTO": {
            "type": "object",
            "properties": {
                "check_out_at": {
                    "type": "string"
                },
                "location_lat": {
                    "type": "number"
                },
                "location_long": {
                    "type": "number"
                }
            }
        },
        "attendance_logs.CreateDTO": {
            "type": "object",
            "required": [
                "status",
                "type",
                "user_id"
            ],
            "properties": {
                "absence_note": {
                    "type": "string"
                },
                "absence_reason_id": {
                    "type": "string"
                },
                "check_in_at": {
                    "type": "string"
                },
                "check_in_lat": {
                    "type": "number"
                },
                "check_in_long": {
                    "type": "number"
                },
                "check_out_at": {
                    "type": "string"
                },
                "check_out_lat": {
                    "type": "number"
                },
                "check_out_long": {
                    "type": "number"
                },
                "learning_group_id": {
                    "type": "string"
                },
                "requires_check_out": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "on_time",
                        "late",
                        "absent"
                    ]
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "student",
                        "teacher",
                        "learning_group"
                    ]
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "attendance_logs.UpdateDTO": {
            "type": "object",
            "properties": {
                "absence_note": {
                    "type": "string"
                },
                "absence_reason_id": {
                    "type": "string"
                },
                "check_in_at": {
                    "type": "string"
                },
                "check_in_lat": {
                    "type": "number"
                },
                "check_in_long": {
                    "type": "number"
                },
                "check_out_at": {
                    "type": "string"
                },
                "check_out_lat": {
                    "type": "number"
                },
                "check_out_long": {
                    "type": "number"
                },
                "learning_group_id": {
                    "type": "string"
                },
                "requires_check_out": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "on_time",
                        "late",
                        "absent"
                    ]
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "student",
                        "teacher",
                        "learning_group"
                    ]
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "auth.LoginDTO": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "remember_me": {
                    "type": "boolean"
                }
            }
        },
        "auth.RegisterDTO": {
            "type": "object",
            "required": [
                "email",
                "institution_name",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "institution_name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "website": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
        "auth.ResetPasswordDTO": {
            "type": "object",
            "required": [
                "password",
                "user_id"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "reset_by_name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "auth.StudentScanLoginDTO": {
            "type": "object",
            "required": [
                "barcode",
                "institution_code",
                "pin"
            ],
            "properties": {
                "barcode": {
                    "type": "string",
                    "maxLength": 255
                },
                "institution_code": {
                    "type": "string",
                    "maxLength": 100
                },
                "pin": {
                    "type": "string",
                    "maxLength": 8,
                    "minLength": 4
                }
            }
        },
        "export.ColumnDTO": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                },
                "foreignKey": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "table": {
                    "type": "string"
                }
            }
        },
        "export.ExportDTO": {
            "type": "object",
            "properties": {
                "column": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/export.ColumnDTO"
                    }
                },
                "filename": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer"
                },
                "models": {
                    "type": "string"
                },
                "pageSize": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}